# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ToNote App OpenApi API Documentation",
        "description": "ToNote App Using L5 Swagger OpenApi description",
        "contact": {
            "email": "schneidershades@gmail.com"
        },
        "version": "3.0"
    },
    "servers": [
        {
            "url": "http://getwordprime.herokuapp.com/",
            "description": "Staging API server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all users",
                "description": "Get all users",
                "operationId": "allUsers",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Post users",
                "description": "Post users",
                "operationId": "postUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/users/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Show user",
                "description": "Show user",
                "operationId": "showUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update user",
                "description": "Update user",
                "operationId": "UserUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete user",
                "description": "Delete user",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserCreateFormRequest": {
                "title": "User Create Form Request Fields",
                "description": "User Create request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "first_name": {
                        "title": "First Name",
                        "description": "first name of the user",
                        "type": "string",
                        "example": "Schneider"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "description": "last name of the user",
                        "type": "string",
                        "example": "Schneider"
                    },
                    "role": {
                        "title": "User Role",
                        "description": "User/Admin",
                        "type": "string",
                        "example": "User"
                    },
                    "email": {
                        "title": "User email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "info@convertscript.com"
                    },
                    "password": {
                        "title": "User password",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "info@convertscript.com"
                    },
                    "plans": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "plan_id": {
                                    "type": "integer",
                                    "example": "1"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "UserUpdateFormRequest": {
                "title": "User Update Form Request Fields",
                "description": "User Update request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "first_name": {
                        "title": "First Name",
                        "description": "first name of the user",
                        "type": "string",
                        "example": "Schneider"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "description": "last name of the user",
                        "type": "string",
                        "example": "Schneider"
                    },
                    "role": {
                        "title": "User Role",
                        "description": "User/Admin",
                        "type": "string",
                        "example": "User"
                    },
                    "email": {
                        "title": "User email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "info@convertscript.com"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}